% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_sites.R
\name{alternative_sites}
\alias{alternative_sites}
\title{Alternative Sampling Site Selection}
\description{
Functions for finding environmentally similar alternative sites when
original sampling locations become inaccessible, based on multivariate
distance in covariate space.
}
\keyword{internal}
