% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.add_samples_to_plot}
\alias{.add_samples_to_plot}
\title{Add Samples to Plot (Internal)}
\usage{
.add_samples_to_plot(p, samples, color, alpha, size)
}
\arguments{
\item{p}{A ggplot object}

\item{samples}{An ss_samples object}

\item{color}{Point color}

\item{alpha}{Point transparency}

\item{size}{Point size}
}
\value{
Modified ggplot object
}
\description{
Add Samples to Plot (Internal)
}
\keyword{internal}
