% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxvol_internal.R
\name{.apply_distance_constraint}
\alias{.apply_distance_constraint}
\title{Apply Distance Constraint During Iteration}
\usage{
.apply_distance_constraint(weights, coords, selected, min_dist)
}
\arguments{
\item{weights}{Numeric vector of selection weights}

\item{coords}{Matrix (m x 2) of coordinates}

\item{selected}{Integer vector of currently selected indices}

\item{min_dist}{Minimum allowed distance}
}
\value{
Modified weights vector
}
\description{
Zeros out the selection weights for candidate points that are too close
to already selected points.
}
\keyword{internal}
