% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxvol_internal.R
\name{.apply_distance_constraint_init}
\alias{.apply_distance_constraint_init}
\title{Apply Distance Constraint to Initial Selection}
\usage{
.apply_distance_constraint_init(coords, k, min_dist)
}
\arguments{
\item{coords}{Matrix (m x 2) of coordinates}

\item{k}{Number of points to select}

\item{min_dist}{Minimum distance between points}
}
\value{
Integer vector of k indices
}
\description{
Selects initial k points with minimum distance constraints.
}
\keyword{internal}
