% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_internal.R
\name{.compute_centroid}
\alias{.compute_centroid}
\title{Compute Cluster Centroid}
\usage{
.compute_centroid(coords, is_latlon = FALSE)
}
\arguments{
\item{coords}{Matrix of coordinates for points in the cluster}

\item{is_latlon}{Logical, whether coordinates are lat/lon}
}
\value{
A vector of length 2 with the centroid coordinates
}
\description{
Computes the centroid of points in a cluster.
}
\keyword{internal}
