% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_optimization.R
\name{.extract_sample_values}
\alias{.extract_sample_values}
\title{Extract Raster Values at Sample Locations}
\usage{
.extract_sample_values(rasters, sample_points)
}
\arguments{
\item{rasters}{A \code{SpatRaster} stack of ancillary variables.}

\item{sample_points}{Spatial points, either an \code{sf} object or a matrix
of x, y coordinates.}
}
\value{
A data frame of extracted values.
}
\description{
Extract Raster Values at Sample Locations
}
\keyword{internal}
