% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_internal.R
\name{.haversine_distance_matrix}
\alias{.haversine_distance_matrix}
\title{Squared Distance Matrix for Lat/Lon Coordinates}
\usage{
.haversine_distance_matrix(X, Y)
}
\arguments{
\item{X}{A matrix of coordinates (n x 2), columns are lon, lat}

\item{Y}{A matrix of coordinates (m x 2), columns are lon, lat}
}
\value{
A matrix of squared distances (n x m) in km^2
}
\description{
Squared Distance Matrix for Lat/Lon Coordinates
}
\keyword{internal}
