% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_internal.R
\name{.haversine_distance_sq}
\alias{.haversine_distance_sq}
\title{Compute Squared Haversine Distance}
\usage{
.haversine_distance_sq(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1, lat1}{Longitude and latitude of first point(s) in degrees}

\item{lon2, lat2}{Longitude and latitude of second point(s) in degrees}
}
\value{
Squared distance in km^2 (approximate)
}
\description{
Computes squared great circle distance for lat/lon coordinates.
}
\keyword{internal}
