% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_internal.R
\name{.kmeans_coverage}
\alias{.kmeans_coverage}
\title{Run K-means with Multiple Tries}
\usage{
.kmeans_coverage(
  cell_coords,
  n_strata,
  prior_coords = NULL,
  max_iter = 1000L,
  n_try = 1L,
  equal_area = FALSE,
  is_latlon = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{cell_coords}{Matrix of cell center coordinates (n x 2)}

\item{n_strata}{Number of strata (clusters) to create}

\item{prior_coords}{Optional matrix of prior point coordinates}

\item{max_iter}{Maximum number of iterations per try}

\item{n_try}{Number of random initializations to try}

\item{equal_area}{Logical, whether to use swop algorithm}

\item{is_latlon}{Logical, whether coordinates are lat/lon}

\item{verbose}{Logical, whether to print progress}
}
\value{
A list with the best clustering result
}
\description{
Runs the k-means algorithm multiple times and returns the best result.
}
\keyword{internal}
