% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_internal.R
\name{.kmeans_swop}
\alias{.kmeans_swop}
\title{K-means Swop Algorithm (Pure R)}
\usage{
.kmeans_swop(cell_coords, n_strata, max_iter = 1000L, is_latlon = FALSE)
}
\arguments{
\item{cell_coords}{Matrix of cell center coordinates (n x 2)}

\item{n_strata}{Number of strata (clusters) to create}

\item{max_iter}{Maximum number of iterations}

\item{is_latlon}{Logical, whether coordinates are lat/lon}
}
\value{
A list with clustering results
}
\description{
Implements the k-means swop algorithm for equal-area strata.
}
\keyword{internal}
