% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_internal.R
\name{.kmeans_transfer}
\alias{.kmeans_transfer}
\title{K-means Transfer Algorithm (Pure R)}
\usage{
.kmeans_transfer(
  cell_coords,
  n_strata,
  prior_coords = NULL,
  max_iter = 1000L,
  is_latlon = FALSE
)
}
\arguments{
\item{cell_coords}{Matrix of cell center coordinates (n x 2)}

\item{n_strata}{Number of strata (clusters) to create}

\item{prior_coords}{Optional matrix of prior point coordinates}

\item{max_iter}{Maximum number of iterations}

\item{is_latlon}{Logical, whether coordinates are lat/lon}
}
\value{
A list with clustering results
}
\description{
Implements the k-means transfer algorithm for spatial coverage sampling.
}
\keyword{internal}
