% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_optimization.R
\name{.plot_kl_cdf}
\alias{.plot_kl_cdf}
\title{Plot Cumulative Density Function Used to Pick the Optimal Sample Size}
\usage{
.plot_kl_cdf(
  sample_sizes,
  cdf_values,
  probability_threshold,
  optimal_sample_size,
  step_size
)
}
\arguments{
\item{sample_sizes}{Numeric vector of tested sample sizes.}

\item{cdf_values}{Numeric vector, CDF of (1 - KL divergence).}

\item{probability_threshold}{Numeric, CDF threshold.}

\item{optimal_sample_size}{Integer, the selected sample size.}

\item{step_size}{Integer, used to position the annotation.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Cumulative Density Function Used to Pick the Optimal Sample Size
}
\keyword{internal}
