% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_internal.R
\name{.squared_distance_matrix}
\alias{.squared_distance_matrix}
\title{Squared Euclidean Distance Matrix}
\usage{
.squared_distance_matrix(X, Y)
}
\arguments{
\item{X}{A matrix of coordinates (n x 2)}

\item{Y}{A matrix of coordinates (m x 2)}
}
\value{
A matrix of squared distances (n x m)
}
\description{
Computes the squared Euclidean distance between each row of X and each row of Y.
}
\keyword{internal}
