% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_optimization.R
\name{kl_optimization}
\alias{kl_optimization}
\title{KL Divergence Sample Size Optimization}
\description{
Functions supporting sample size optimization for cLHS designs based on
Kullback-Leibler divergence between population and sample distributions,
following Malone et al. (2019).
}
\keyword{internal}
