% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ss_samples}
\alias{plot.ss_samples}
\title{Plot Method for ss_samples Objects}
\usage{
\method{plot}{ss_samples}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ss_samples}}

\item{...}{Additional arguments passed to \code{\link[=ss_plot]{ss_plot()}} or \code{\link[=ss_plot_samples]{ss_plot_samples()}}}
}
\value{
A ggplot object
}
\description{
Plot Method for ss_samples Objects
}
