% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ss_strata}
\alias{plot.ss_strata}
\title{Plot Method for ss_strata Objects}
\usage{
\method{plot}{ss_strata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ss_strata}}

\item{...}{Additional arguments passed to \code{\link[=ss_plot]{ss_plot()}}}
}
\value{
A ggplot object
}
\description{
Plot Method for ss_strata Objects
}
