% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{ss_area}
\alias{ss_area}
\title{Get Area of Each Stratum}
\usage{
ss_area(x)
}
\arguments{
\item{x}{An object of class \code{ss_strata}}
}
\value{
Named numeric vector with area of each stratum
}
\description{
Get Area of Each Stratum
}
