% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{ss_coverage_equal_area}
\alias{ss_coverage_equal_area}
\title{Spatial Coverage Sampling with Equal-Area Strata}
\usage{
ss_coverage_equal_area(
  x,
  n_strata,
  n_try = 1L,
  n_cells = 2500L,
  verbose = FALSE
)
}
\arguments{
\item{x}{An \code{sf} object representing the study area.}

\item{n_strata}{Integer, number of strata to create.}

\item{n_try}{Integer, number of random initializations (default 1).}

\item{n_cells}{Integer, approximate number of grid cells (default 2500).}

\item{verbose}{Logical, whether to print progress (default FALSE).}
}
\value{
An object of class \code{ss_samples}.
}
\description{
Generates sampling points using spatial coverage with equal-area strata.
Useful when strata need to have similar sizes.
}
\details{
This function creates strata of approximately equal area using the
swopping algorithm, then returns centroid samples.
}
\examples{
\dontrun{
library(sf)
poly <- st_polygon(list(rbind(c(0, 0), c(100, 0), c(100, 50), c(0, 50), c(0, 0))))
study_area <- st_sf(geometry = st_sfc(poly))

samples <- ss_coverage_equal_area(study_area, n_strata = 25)
}

}
\seealso{
\code{\link[=ss_coverage]{ss_coverage()}}, \code{\link[=ss_composite]{ss_composite()}}
}
