% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{ss_get_samples}
\alias{ss_get_samples}
\title{Get Samples as sf Object}
\usage{
ss_get_samples(x)
}
\arguments{
\item{x}{An object of class \code{ss_samples}}
}
\value{
An \code{sf} object with the sampling points
}
\description{
Get Samples as sf Object
}
