% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_optimization.R
\name{ss_load_rasters}
\alias{ss_load_rasters}
\title{Load Predictor Rasters}
\usage{
ss_load_rasters(path)
}
\arguments{
\item{path}{Character, path to a \code{.tif} file or a directory containing
\code{.tif} files.}
}
\value{
A \code{SpatRaster} stack.
}
\description{
Loads either a multi-layer raster file or a directory of single-layer
raster files into a \code{SpatRaster} stack.
}
\examples{
\dontrun{
predictors <- ss_load_rasters("data/predictors.tif")
}

}
