% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{ss_n_samples}
\alias{ss_n_samples}
\title{Get Number of Samples}
\usage{
ss_n_samples(x)
}
\arguments{
\item{x}{An object of class \code{ss_samples}}
}
\value{
Integer, the number of samples
}
\description{
Get Number of Samples
}
