% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{ss_n_strata}
\alias{ss_n_strata}
\title{Get Number of Strata}
\usage{
ss_n_strata(x)
}
\arguments{
\item{x}{An object of class \code{ss_strata}}
}
\value{
Integer, the number of strata
}
\description{
Get Number of Strata
}
