% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ss_plot_samples}
\alias{ss_plot_samples}
\title{Plot Sampling Points}
\usage{
ss_plot_samples(x, color = "black", alpha = 0.7, size = 2, ...)
}
\arguments{
\item{x}{An object of class \code{ss_samples}.}

\item{color}{Character, point color (default "black").}

\item{alpha}{Numeric, point transparency (default 0.7).}

\item{size}{Numeric, point size (default 2).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a simple plot of sampling points.
}
\examples{
\dontrun{
library(sf)
poly <- st_polygon(list(rbind(c(0, 0), c(100, 0), c(100, 50), c(0, 50), c(0, 0))))
study_area <- st_sf(geometry = st_sfc(poly))

samples <- ss_random(study_area, n = 30)
ss_plot_samples(samples)
}

}
\seealso{
\code{\link[=ss_plot]{ss_plot()}}
}
