% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{ss_relative_area}
\alias{ss_relative_area}
\title{Get Relative Area of Each Stratum}
\usage{
ss_relative_area(x)
}
\arguments{
\item{x}{An object of class \code{ss_strata}}
}
\value{
Named numeric vector with relative area (proportion) of each stratum
}
\description{
Get Relative Area of Each Stratum
}
