% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxvol.R
\name{ss_summary.maxvol}
\alias{ss_summary.maxvol}
\title{Summary Method for Maxvol Sampling Results}
\usage{
ss_summary.maxvol(object, ...)
}
\arguments{
\item{object}{An object of class \code{ss_samples} from \code{ss_maxvol()}}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{object}, invisibly. Called for the side effect of printing a
maxvol design summary to the console. If \code{object} was not produced by
\code{\link[=ss_maxvol]{ss_maxvol()}}, dispatch is passed on with \code{NextMethod()} and that
method's value is returned instead.
}
\description{
Summary Method for Maxvol Sampling Results
}
