% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ss_summary}
\alias{ss_summary}
\title{Create Summary Statistics Table}
\usage{
ss_summary(x)
}
\arguments{
\item{x}{An object of class \code{ss_samples}}
}
\value{
A data frame with summary statistics
}
\description{
Creates a summary of the sampling design.
}
