% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{ss_to_data_frame}
\alias{ss_to_data_frame}
\title{Convert to Data Frame}
\usage{
ss_to_data_frame(x, what = "samples")
}
\arguments{
\item{x}{An object of class \code{ss_samples} or \code{ss_strata}}

\item{what}{Character, what to extract (default "samples")}
}
\value{
A data frame with coordinates
}
\description{
Convert to Data Frame
}
