% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{ss_to_sf}
\alias{ss_to_sf}
\title{Convert to sf Object}
\usage{
ss_to_sf(x, what = "samples")
}
\arguments{
\item{x}{An object of class \code{ss_samples} or \code{ss_strata}}

\item{what}{Character, what to extract: "samples", "cells", or "centroids"}
}
\value{
An \code{sf} object
}
\description{
Convert to sf Object
}
