% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{summary.ss_samples}
\alias{summary.ss_samples}
\title{Summary Method for ss_samples Objects}
\usage{
\method{summary}{ss_samples}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ss_samples}}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{object}, invisibly. Called for the side effect of printing a
detailed summary of the sampling design to the console.
}
\description{
Summary Method for ss_samples Objects
}
