% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{summary.ss_strata}
\alias{summary.ss_strata}
\title{Summary Method for ss_strata Objects}
\usage{
\method{summary}{ss_strata}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ss_strata}}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{object}, invisibly. Called for the side effect of printing a
detailed summary of the stratification to the console.
}
\description{
Summary Method for ss_strata Objects
}
