% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.ss_coverage_efficiency}
\alias{print.ss_coverage_efficiency}
\title{Print Method for ss_coverage_efficiency Objects}
\usage{
\method{print}{ss_coverage_efficiency}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ss_coverage_efficiency}}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for ss_coverage_efficiency Objects
}
