% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{print.ss_samples}
\alias{print.ss_samples}
\title{Print Method for ss_samples Objects}
\usage{
\method{print}{ss_samples}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ss_samples}}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for ss_samples Objects
}
