% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{print.ss_strata}
\alias{print.ss_strata}
\title{Print Method for ss_strata Objects}
\usage{
\method{print}{ss_strata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ss_strata}}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for ss_strata Objects
}
