% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxvol.R
\name{ss_summary.maxvol}
\alias{ss_summary.maxvol}
\title{Summary Method for Maxvol Sampling Results}
\usage{
ss_summary.maxvol(object, ...)
}
\arguments{
\item{object}{An object of class \code{ss_samples} from \code{ss_maxvol()}}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for Maxvol Sampling Results
}
